Add RoaringBitmapArray as replacement for Roaring64Bitmap
Adds a new bitmap implementation called `RoaringBitmapArray`, which replaces usage
of `Roaring64Bitmap` in deletion vectors. This implementation is optimized for use
case of handling row indices, which are always clustered between 0 and the index
of the last row number of a file, as opposed to being arbitrarily sparse over the
whole `Long` space. The implementation is much simpler than
`Roaring64Bitmap` and less error prone, as well as faster.

GitOrigin-RevId: 862989348614520672065786c5607f0ade7a93e7