Add utilities for encoding/decoding of Deletion Vectors bitmap
Deletion vectors are stored either in a separate file or inline in the `AddFile` struct in DeltaLog.
More details of the format are found here: https://github.com/delta-io/delta/blob/master/PROTOCOL.md#deletion-vector-descriptor-schema

In this PR, add utilities to encode/decode the DV bitmap in Base85 for storing it in `AddFile` struct in DeltaLog.

GitOrigin-RevId: 9351b556fcf1d25ef864cd98406128c684e4b379