Add RoaringBitmapArray to store indices of rows deleted in a data file
This PR is part of the feature: Support reading Delta tables with deletion vectors (more at delta-io/delta#1485)

Adds a new bitmap implementation called `RoaringBitmapArray`. This will be used to encode the deleted row indices. There already exists a `Roaring64Bitmap` provided by the `org.roaringbitmap` library , but this implementation is optimized for use case of handling row indices, which are always clustered between 0 and the index of the last row number of a file, as opposed to being arbitrarily sparse over the whole `Long` space.

Closes delta-io/delta#1486

GitOrigin-RevId: c94d0cd1b4d1c179b9947e984705ab81e97a6dec