Add check to enforce at most one Protocol action per commit
This PR adds a safety check to ensure that every commit to a Delta table contains at most one Protocol action. As the added UT demonstrates, it is possible to have more than one Protocol action in a commit when manually constructing a commit. This can be a problem if Protocols conflict because Delta will just pick one of the two version when determining the current table version. This can lead to error because features that should be enabled on a table can be rejected.

Currently, commits generated by Delta should be safe from this issue as none of the existing code paths can end up producing more than one protocol version. However, we are adding this additional check to future proof the code so that if such a case ever occurs (either through a newly introduced code path or by manually constructing a commit), it is caught early.

GitOrigin-RevId: 7f5318c7c0606495fce8410572f1b97f2610f893