Update Spec to use the portable Deletion Vector serialization format
## Description

The RoaringBitmap community has recently standardised on a ["portable" 64-bit serialization format](https://github.com/RoaringBitmap/RoaringFormatSpec#extention-for-64-bit-implementations).
We should switch the Deletion Vector feature to use this format, as it will make adoption by 3rd parties easier, enabling them to use an off-the-shelf 64-bit implementation from the RoaringBitmap library of their choice, as long as it supports this format (which CRoaring already does and the next release of Java RoaringBitmap will do as well).
This PR makes the appropriate changes to the DV format description in the appendix of the Delta spec.

N/A

## Does this PR introduce _any_ user-facing changes?

No.

Closes delta-io/delta#1494

Signed-off-by: Venki Korukanti <venki.korukanti@gmail.com>
GitOrigin-RevId: 1503f42337a553879ad527bee8b78bf2d034bfd2