Fix the vacuum log message in DRY RUN
Currently, we output a log line with `DeltaVacuumStats` like this
```
Found 12 files and directories in a total of 1 directories that are safe to delete.DeltaVacuumStats(true,Some(0),604800000,1669825818524,1,12,10404,11374,0)
```

This change fixes it in a better format:
```
Found 12 files (10404 bytes) and directories in a total of 1 directories that are safe to delete.
```

A simple log change.

GitOrigin-RevId: 182e7c094647f639e925d714f60f2b4dc8c32944