Support SHALLOW CLONE for Delta Lake
This PR introduces support for **SHALLOW CLONE** for both Delta **and Parquet table** on Delta Lake, specifically the following command:
`CREATE [OR REPLACE] TABLE [IF NOT EXISTS] target SHALLOW CLONE source [VERSION AS OF version | TIMESTAMP AS OF timestamp] [TBLPROPERTIES clause] [LOCATION path]`

This enables the following use cases:
1. Create a `target` table with Delta log pointing to the files from the `source` table. The source table can be either a Delta table or a Parquet table.

You may also specify a custom `path` to create as an external table in a path, a `clause` for additional table properties to append to, or a `version` to create the target table as a time-travelled version of the source table (if is Delta).

E.g. `CREATE TABLE target SHALLOW CLONE source`

2. Replace/restore a table by shallow-cloning itself. This requires the table to be **empty** by the time of cloning to avoid data duplication.

E.g. `REPLACE TABLE source SHALLOW CLONE source VERSION AS OF 1`

Closes https://github.com/delta-io/delta/pull/1505

Existing tests.

GitOrigin-RevId: 6b609eddc8c49c2831bd1d7a77660ae817c6246e