Add `DeletionVectorDescriptor` to represent DV metadata in `DeltaLog`
This PR is part of the feature: Support reading Delta tables with deletion vectors (more details of supporting DV table reads at delta-io/delta#1485)

This PR adds a new class called `DeletionVectorDescriptor` which is used to represent the deletion vector for a given file (`AddFile`) in `DeltaLog`. The format of the metadata is described in the [protocol](https://github.com/delta-io/delta/blob/master/PROTOCOL.md#deletion-vector-descriptor-schema).

GitOrigin-RevId: 59979dea64d7c938d33b886eaebe0b30b2f8dd6c