Remove check when adding feature descriptors to a protocol
This PR fixes a bug when the following method is called twice in a row:
```scala
var p = Protocol(3, 7)
// Protocol(3, 7, {}, {})

p = p.withFeatureDescriptor(feature, addToReaderFeatures = true, addToWriterFeatures = false)
// Return: Protocol(3, 7, {feature}, {})

p = p.withFeatureDescriptor(feature, addToReaderFeatures = false, addToWriterFeatures = true)
// Return: Protocol(3, 7, {feature}, {})
// Expected: Protocol(3, 7, {feature}, {feature})

```
The issue is the second call will not add `feature` to `writerFeratures`. It's because the check sees it already exists in `readerFeatures`, therefore will not try to add it to `writerFeratures`.

This PR removes that check.

GitOrigin-RevId: 8b14c5e91c9e610c142d0bfe613563f49d95c9c8