Eliminate most calls to the unsafe DeltaLog.metadata method
Because it relies on `DeltaLog.unsafeVolatileSnapshot`, which can change at any time without notice as other threads access the Delta log, the `DeltaLog.metadata` method is also unsafe. This PR eliminates almost all call sites of the method, and renames it `DeltaLog.unsafeVolatileMetadata` to make the danger more clear to callers.

GitOrigin-RevId: 2e0a5b4b211c5b010a188adf08ff949a43c2e0ec