Port existing features to Table Features
This PR adapts the following legacy features to Table Features:

- appendOnly
- invariants
- checkConstriants
- changeDataFeed
- generatedColumns
- columnMapping
- identityColumns

Note that Deletion Vector will be ported in a separate PR.

This PR does not modify each feature to check the `protocol` action to determine if it's used. Instead, there's a one-time check (in `Snapshot`) when opening a table for read, to ensure all legacy features implicitly enabled in metadata are referenced in `protocol`.

Closes delta-io/delta#1530

GitOrigin-RevId: 0e89a718271d3ae94efce2bf2aef255beaad4d96