Remove feature descriptors in Table Features
This change addresses feedback in https://github.com/delta-io/delta/pull/1450#discussion_r1054717897 to remove all uses of `TableFeatureDescriptor` in the codebase because that `status` field in the descriptor can only be `enabled` which can be omitted. Therefore we could simplify store feature names as `Set[String]` in the protocol and assume all listed features are `enabled`.

`Protocol` action before this change:
```json
{
  "protocol":{
    "readerVersion":3,
    "writerVersion":7,
    "readerFeatures":[
      {"name":"columnMapping","status":"enabled"}
    ],
    "writerFeatures":[
      {"name":"columnMapping","status":"enabled"},
      {"name":"identityColumns","status":"enabled"}
    ]
  }
}
```
`Protocol` action after this change:
```json
{
  "protocol":{
    "readerVersion":3,
    "writerVersion":7,
    "readerFeatures":["columnMapping"],
    "writerFeatures":["columnMapping","identityColumns"]
  }
}
```

GitOrigin-RevId: 445601bbaebf39f1cd41f8257ece7aeaff48e101