Legacy table features enabled via metadata can be enabled
When we enable a table feature from metadata only, we figure out the minimum required protocol to accommodate the metadata. For legacy features, this means they could end up in the implicitly enabled features, even though they have been enabled explicitly via the metadata if the minimum required protocol to accommodate the metadata is below activating table features.

While initializing the protocol, we might bump the protocol version so that table features are enabled. If table features are enabled, this means there are no implicit enabled table features anymore. However, this means we lose the implicit table features that have been enabled via the metadata, as they are not carried over.

To combat this, we first have to determine the protocol version, and then use this to add table features that are enabled by metadata either to the implicitly enabled features or the set of enabled table features.

GitOrigin-RevId: 8c8f84e6d3951f09bdb336bcd9ab59fbd7368099