Register VACUUM operations in the delta log
This PR registers the start and end of VACUUM operations in the delta log. This means that we commit a commit with no Add/Remove files, and only a `CommitInfo` file which contains the delta operation info.

`VacuumStart` operation contains metrics: `numFilesToDelete` and `sizeOfDataToDelete`

`VacuumEnd` operation contains metrics: `numDeletedFiles` and `numVacuumedDirectories`

New UTs.

Expose additional metrics and history in the _delta_log for the start and end of VACUUM operations.

Closes delta-io/delta#1552.
Resolves delta-io/delta#868.

Co-authored-by: Yann Byron <biyan900116@gmail.com>
GitOrigin-RevId: 94805531d022bac4afafd0b672d17b8828d8aa2c