Don't drop table feature set by table property override during CLONE
In CLONE, we merged the `txn.metadata` to come up with a right protocol version to account for table property overrides. With table features, this approach does not work anymore as it doesn't account for table features set in the table property overrides, as they end up in the protocol of the transaction, and not the metadata.

This is fixed by merging the table property overrides with the `txn.metadata` when coming up with the minimal protocol components.

GitOrigin-RevId: 18aaf85df24b80970b5098d6e891451c875ea79c