Forbid protocol downgrades when replacing Delta tables
Replacing a Delta table is implemented as committing a new Delta version that clears existing data and updates metadata. Currently this is treated as creating a new table and the protocol version is determined based on existing settings and used features.  However, the the table's history is not cleared and customers can time travel or read CDF from previous versions. If the new version results in a protocol downgrade, these operations may return wrong data.

This commit fixes this issue by preventing protocol downgrades when replacing a table. Replacing a table is treated as
a commit that is allowed to reset all metadata, but only upgrade the protocol.

Note, that this change has two important side-effects:
* Replacing a table will no longer pick the latest default table protocol.
* Users can no longer perform an in-place downgrade using `REPLACE TABLE X AS SELECT * FROM X`. This was allowed until now, but could return wrong results for time-travel/cdf queries.

GitOrigin-RevId: a2aa81835981cd7b502e6df1b755153889cfb878