Support limit pushdown on Delta tables with DVs.
This PR is part of the feature: Support reading Delta tables with deletion vectors (more details at https://github.com/delta-io/delta/issues/1485)

Currently the limit pushdown code doesn't take into account the DVs when pruning the list of files based on the `limit` value. Update the code to take `dv.cardinality` when finding the number of rows in a `AddFile`.

It also adds
 * comments around data skipping code and how data skipping continue to work without giving wrong results (with a bit performance overhead) when querying tables with DVs
 * test utilities to associate DVs with `AddFile` for testing purposes.

Closes delta-io/delta#1577

GitOrigin-RevId: fdb0c78db9b2d0ac37cce67886110a32688fd531