Generate update expressions for new nested target fields in PreprocessTableMerge
With schema evolution, new nested columns can be added to the target table during a merge operation. The code used to generate expression to set these columns to null when they are not otherwise set by a merge clause only handles top-level columns. It is extended here to also handle evolution of nested attributes.

GitOrigin-RevId: e0592fdc6ce6c3b98a5153f48b65ff4b1d905499