Fix a bug where selecting `_metadata` columns on Delta tables with DVs fails
Currently when rewriting the scan `PreprocessTablesWithDVs`, we use the `Scan` relation output as the data output for `HadoopFsRelation`. Data output comes directly from the data file or data file reader. Scan output contains additional columns such as `_metadata` which are populated in Scan operator once the data is read from the file reader.

Added a test. Also added a golden table for partitioned tables with DVs and tests.

Closes delta-io/delta#1583

Signed-off-by: Venki Korukanti <venki.korukanti@gmail.com>
GitOrigin-RevId: ee0d1d43e38597b653fdd5f47163b32710393054