Block manifest file generation when DVs are present
This PR is part of the feature: Support reading Delta tables with deletion vectors (more details at https://github.com/delta-io/delta/issues/1485)

Manifest file contains the list of Parquet data files paths that can be consumed by clients with symlink table type reading capability. However when DVs are present on top of the Parquet data files, there is no way to expose the DVs to symlink table reader. It is best to block manifest file generation when DVs are present in the table.

Closes delta-io/delta#1595

GitOrigin-RevId: 2589bcbd08c59bb65c5dc465d31b5e0e4aabf5c0