[DELETE with DVs] Update `DeltaParquetFileFormat` to return a row index
This PR is part of [DELETE with Deletion Vector implementation](https://github.com/delta-io/delta/pull/1591).

It updates the `DeltaParquetFileFormat` to populate row index when the schema contains the columns. In order to generate the row_index, we disable file splitting and filter pushdown into reader. This is temporary until we upgrade to Parquet reader in Spark 3.4 that generates the row_indexes with file splitting and filter pushdown.

Closes delta-io/delta#1593

Signed-off-by: Venki Korukanti <venki.korukanti@gmail.com>
GitOrigin-RevId: 2a3cf2816eff576346ad3f0edebd3e09cd86f780