Return table name in DeltaTable.detail()
`DeltaTable.detail()` API currently does not return the name of the Delta tables. This is because `DescribeDeltaDetailCommand.getPathAndTableMetadata()` skips the catalog lookup when a path is set, assuming that either a path or a table identifier will be provided. This commit fixes this issue by always performing a lookup to the catalog when a table identifier is provided.

Closes delta-io/delta#1549

GitOrigin-RevId: a899637cd384e6e303809ac3106b9159bcfae0e4