Removing config commit info enabled
DELTA_COMMIT_INFO_ENABLED is an old config that is already enabled by default. This PR removes the config, as there are no reasons for a customer to explicitly disable the config. This allows for lesser dependencies on DELTA_USER_METADATA and DELTA_VACUUM_LOGGING_ENABLED which currently rely on DELTA_COMMIT_INFO_ENABLED to be enabled.

GitOrigin-RevId: 95f24cebe5c518a5f4d55bbd1e1852fb764b9bc1