Get rid of previousOffset in DeltaSource
DeltaSource was storing its latest processed offset in `previousOffset`, in `getBatch`. At the same time, it was ignoring the `startingOffset` that is passed into `latestOffset`, which is the parameter that is actually intended to convey this information. This PR gets rid of `previousOffset` and actually uses the `latestOffset` parameter for this.

In addition, the PR changes the `DeltaSource` code to use `DeltaSourceOffset` consistently inside the internals of the class, instead of passing the base class `Offset` everywhere and then converting back to `DeltaSourceOffset` all over the place. Furthermore, it adds a utility function to do the conversion from `Offset` to `DeltaSourceOffset`.

GitOrigin-RevId: b47de2fe9ce3381c09b4c5efd15ecb828c3f39e2