Add test coverage for special characters in Delta table path
This change provides coverage for special characters in Delta table path in all tests by default by including a space in the default temp directory used when creating test tables in DML tests.
An improved helper to parse the table name and table alias from tests that behaves correctly when the table name or path contains spaces is provided.

GitOrigin-RevId: 989649fd59ad33e55400786fb43924b6d8040480