Adds DeltaTableV2.tableExists helper method
This PR is a minor refactor that adds a tableExists helper method to DeltaTableV2. This allows callers to directly call tableExists on a DeltaTableV2 instead of going through the DeltaTableV2's deltaLog member (which is currently required). Instead of `table.deltaLog.tableExists`, callers can now just call `table.tableExists` (where `table` is a DeltaTableV2).

GitOrigin-RevId: f07c3977733b9df9f5087318a13526c30465ab4c