Ignore internal column mapping table properties when creating table
When creating a table against an existing Delta location, a comparison is made between the configuration of the newly created table and the existing Delta table. The create fails if this comparison fails. However, some internal column mapping properties cannot be set by the create so they should be ignored for the purposes of this verification.

GitOrigin-RevId: 29505a33c0c99678b50448f048fbdf5ef8c20a7d