Block writing to Delta tables that supports identity columns
This PR fixes an issue described in https://github.com/delta-io/delta/issues/1694, where it is possible to INSERT values into an identity column without updating the high watermark.

This issue is caused by a misplaced check in `actions.scala`. The check didn't fire for INSERTs.

Added new tests.

Yes. After this PR is merged, it will no longer be possible to write to a table that has `minWriterVersion` = `6`, or has `identityColumns` in `writerFeatures`.

Closes delta-io/delta#1695

Signed-off-by: Paddy Xu <xupaddy@gmail.com>
GitOrigin-RevId: 048bd370b7c50fa052650c77ba96469cdda5f0e9

# Conflicts:
#	core/src/main/scala/org/apache/spark/sql/delta/TableFeature.scala
#	core/src/main/scala/org/apache/spark/sql/delta/actions/TableFeatureSupport.scala