Adding new stats tracker for auto-stats.
In this PR we are adding a new stats tracker that collects both analyze statistics and Delta statistics. The tracker is added in parallel with the Delta stats tracker and eventually will replace it.

The tracker tasks collect per file analyze statistics during data ingestion. The file level statistics are then aggregated to task level statistics. The tracker job that runs in the driver aggregates the task level statistics to create table level statistics.

GitOrigin-RevId: f7cb4aae152207f09e2bb6f6acc03282a5ffa4b3