Enable SQL support for WHEN NOT MATCHED BY SOURCE
# Description
The SQL syntax for merge with WHEN NOT MATCHED BY SOURCE clauses was shipped with Spark 3.4. Now that Delta picked up Spark 3.4, we can enable SQL support and mix in SQL tests for WHEN NOT MATCHED BY SOURCE.

Existing tests for WHEN NOT MATCHED BY SOURCE are now run in the Merge SQL suite.

Closes delta-io/delta#1740

GitOrigin-RevId: 1ddd1216e13f854901da47896936527618ea4dca