SHALLOW CLONE support for Delta tables with deletion vectors.
This PR is part of the feature: Support Delta tables with deletion vectors (more details at https://github.com/delta-io/delta/issues/1485)

This PR adds support for SHALLOW CLONEing a Delta table with deletion vectors. The main change is to convert the relative path of DV file in `AddFile` to absolute path when cloning the table.

Added tests

Closes delta-io/delta#1733

GitOrigin-RevId: b634496b57b93fc4b7a7cc16e33c200e3a83ba64