DV stress test: Delete from a table of a large number of rows with DVs
This PR tests DELETing from a table of 2 billion rows (`2<<31 + 10`), some of which are marked as deleted by a DV. The goal is to ensure that DV can still be read and manipulated in such a scenario.

We don't `delete a large number of rows` and `materialize DV` because they run too slow to fit in a unit test (9 and 20 minutes respectively).

GitOrigin-RevId: 1273c9372907be0345465c2176a7f76115adbb47