Dependencies between Delta Table Features
This PR adds the ability to specify dependencies between Delta table features. When enabling a table feature all dependent table features will automatically be enabled as well. This will be used by Row Tracking to add a dependency on the Domain Metadata feature (which will be used to store the high water mark).

Closes delta-io/delta#1789

GitOrigin-RevId: c0b03df9d708f1c0ff9b902679a06357567d3907