Rewrite OPTIMIZE to use Spark Table Resolution
This PR is part of https://github.com/delta-io/delta/issues/1705.

It rewrites the Delta OPTIMIZE command to use Spark's table resolution logic instead of resolving the target table manually at command execution time. For that, it changes OptimizeTableCommand from a LeafRunnableCommand to a UnaryLike command that takes a UnresolvedTable as a child plan node, which will be resolved by Spark. In addition, it also introduces a UnresolvedPathBasedDeltaTable LogicalPlan for the case that OPTIMIZE is run against a raw path. This new logical plan is an indication that the target table specified as a path so no table resolution is necessary.

Through the large set of existing unit tests that OPTIMIZE already has + adapted the DeltaSqlParserSuite to work with the new OptimizeTableCommand.

N/A

Fixes https://github.com/delta-io/delta/issues/1705.
Closes https://github.com/delta-io/delta/pull/1708.

GitOrigin-RevId: 33cb734ec9c9082aa29de790f6fc5e5959705d93