Replace UDF to increment metrics in DML commands with native IncrementMetric expression
## Description
This change introduces a new expression `IncrementMetric` to support the frequent use case of incrementing metrics in Merge, Update and Delete. It replaces the use of UDFs to achieve that goal.

Added IncrementMetricSuite to cover usage of the newly introduced expression.
Metrics correctness is covered in existing test suites `MergeIntoMetricsBase`, `UpdateMetricsSuite` and `DeleteMetricsSuite`.

Closes delta-io/delta#1828

GitOrigin-RevId: 746d761b5d2a51651e9d1df0eb1fcae0280791fa