Universal Format (Iceberg) Implementation
This commit contains all of the code needed to implement the UniForm feature in delta-spark.

When enabled, delta-spark will convert Delta metadata to Iceberg metadata asynchronously.

Includes one end-to-end test.

Note that we had to include a few minor modifactions/additions to Iceberg source code, which we do via patch files.