Resolve merge actions in batch
Expressions in merge clauses are resolved one at a time by running a full analysis on a fake plan that includes that single expression. The time required to resolve `UPDATE SET *` or `INSERT *` actions then grows quadratically with the number of columns in the target and source tables respectively: the number of analysis pass grows linearly with the number of expressions, the cost of each pass grows linearly with the number of columns.

This change makes the analysis time linear by resolving all `UPDATE SET *` or `INSERT *` assignment expressions in a single analysis pass, using a fake plan that includes all expressions to resolve.

In addition, the logic to pass the plans that should be used for resolution is simplified. Instead of passing a dummy plan with the right children - either the source only, the target only or both - we directly pass the list of plans to use to the resolution helper.

- Confirmed locally that the analysis time is orders of magnitude faster when the source and table contain 5000 columns.

Closes delta-io/delta#1875

GitOrigin-RevId: 6208cc52aead6a04c80b602ee9d4018c48e9b816