Probe whether the metadata path is canonicalized in Spark
The mechanism of this fix is to call the Spark internal method, which is used to generate metadata columns, to see if it will canonicalize spaces in a crafted path string. If the answer is yes, then we don't need to do anything on the Delta side; otherwise, we manually canonicalize the obtained metadata column.

Why don't use the Spark internal method on `FileToDvDescriptor`, so both sides of the join are either canonicalized or not-canonicalized? Because most Delta methods are expecting a canonicalized path, thus the returned DF must be canonicalized in all cases.

Closes https://github.com/delta-io/delta/pull/1769.
PR targeting `branch-2.4`: https://github.com/delta-io/delta/pull/1770.

GitOrigin-RevId: 3538b18ff23e81c603acbc7df3930ba1730903f2