Re-enable iceberg module and clone iceberg github repo using shallow clone
#### Which Delta project/connector is this regarding?

-Spark
- [ ] Standalone
- [ ] Flink
- [ ] Kernel
- [ ] Other (fill in here)

## Description

This PR uncomments the iceberg module (it was commented out previously to prevent compilation).

This PR also changes how we clone the iceberg src code (in order to be able to apply patch files). Instead of using a deep clone, we use a shallow clone. This had some ramifications in how the JAR files were named ... so needed to update the python generation script, too.

Closes delta-io/delta#1917

Signed-off-by: Scott Sandre <scott.sandre@databricks.com>
GitOrigin-RevId: d8249310313b4a01c387e81e9bc5c6a9d8d0cf0b