Extend remove actions with stats field
Extend the `remove` actions with statistics about the data in the file that is being removed from the table and that can be used to optimize queries that read the changes of older versions, e.g., Change Data Feed queries. These statistics are copied from the `add` action when a file is being removed.

File statistics can be large and increase the size of checkpoints. However, statistics of `remove` actions are not needed for tombstones but only for Delta JSON files, and so are dropped from checkpoints.

Resolves https://github.com/delta-io/delta/pull/1907

GitOrigin-RevId: 5013848522e906fea207bea5078a2e4a0ce605c8