Add more DDL test coverage and support ALTER TABLE REPLACE COLUMNS
## Description

This PR adds more DDL test coverage around commands like SHOW TBLPROPERTIES, SHOW COLUMNS, DESC COLUMN and also adds support for ALTER TABLE REPLACE COLUMNS.

REPLACE COLUMNS as it is implemented today drops all columns and then adds all new columns. This is actually undesirable for a Delta table, as it will cause all data to be dropped when column mapping is enabled. Therefore the way we implement is to look at the final schema, which is provided by all the ADD COLUMN table changes that are provided and try to resolve the differences. If a difference is ambiguous, then we will throw an error.

A column being dropped and added would lead to an ambiguous change, because we don't know if we should be treating that as a column rename or literally a drop + add. Users should explicitly call ALTER TABLE drop columns + ALTER TABLE add columns instead.

This patch introduces many new unit tests

## Does this PR introduce _any_ user-facing changes?

Yes, this PR may start throwing ambiguous change errors for ALTER TABLE REPLACE COLUMNS. I doubt that it is used as much, but it would prevent accidental deletion of data. Users can do a REPLACE TABLE if they really wanted to replace the table with empty data.

Closes delta-io/delta#1822

Co-authored-by: Ryan Johnson <ryan.johnson@databricks.com>
Signed-off-by: Burak Yavuz <brkyvz@gmail.com>
GitOrigin-RevId: e1ff7c1d2c03028fa4803bdc13456fbc63448cc4